fixed the position of the slider component
